# Implement stump_tiles algorithm (#248)