--- a/stumpy/__init__.py
+++ b/stumpy/__init__.py
@@ -2,6 +2,7 @@ from pkg_resources import get_distribution, DistributionNotFound
 import os.path
 from .core import mass  # noqa: F401
 from .stump import stump  # noqa: F401
+from .stump_tiles import stump_tiles  # noqa: F401
 from .stumped import stumped  # noqa: F401
 from .mstump import mstump, subspace, mdl  # noqa: F401
 from .mstumped import mstumped  # noqa: F401