--- a/stumpy/core.py
+++ b/stumpy/core.py
@@ -2933,16 +2933,16 @@ def _get_tiles(m, n_A, n_B, diags, tile_length):
     tiles = np.empty((tiles_count, 7), dtype=np.int64)
 
     for tile_idx in prange(tiles_count):
-        y_offset = (tile_idx // tile_columns) * tile_length
-        x_offset = (tile_idx % tile_columns) * tile_length
+        i_offset = (tile_idx // tile_columns) * tile_length
+        j_offset = (tile_idx % tile_columns) * tile_length
         # height of current tile
-        tile_height = min(tile_length, l - y_offset)
+        tile_height = min(tile_length, l - i_offset)
         # width of current tile
-        tile_width = min(tile_length, w - x_offset)
+        tile_width = min(tile_length, w - j_offset)
 
         # lower and upper diagonal indices to traverse within tile
-        tile_lower_diag = max(1 - tile_height, diags[0] + y_offset - x_offset)
-        tile_upper_diag = min(tile_width, diags[-1] + 1 + y_offset - x_offset)
+        tile_lower_diag = max(1 - tile_height, diags[0] + i_offset - j_offset)
+        tile_upper_diag = min(tile_width, diags[-1] + 1 + i_offset - j_offset)
 
         tile_ndist = _total_diagonal_ndists(
             tile_lower_diag,
@@ -2951,8 +2951,8 @@ def _get_tiles(m, n_A, n_B, diags, tile_length):
             tile_width,
         )
 
-        tiles[tile_idx, 0] = y_offset
-        tiles[tile_idx, 1] = x_offset
+        tiles[tile_idx, 0] = i_offset
+        tiles[tile_idx, 1] = j_offset
         tiles[tile_idx, 2] = tile_height
         tiles[tile_idx, 3] = tile_width
         tiles[tile_idx, 4] = tile_lower_diag