… stumpy/config.py
… test.sh
… tests/test_core.py
… tests/test_stump_tiles.py