This is a pull-request addressing issue #248, i.e. implementation of an algorithm that leverages cache lines to improve the performance of `stump`.

> Note: This is a new and updated version of PR #463, which has been accidentally wiped off, thanks to my stellar version controlling skills :upside_down_face:

# How `stump` works

Let us first briefly explore how `stump` works right now and what's wrong with it.

Consider a time series of length `20` and a window length of `6`. We know that the distance matrix for this configuration will be a `15 x 15` matrix (`20 - 6 + 1 = 15`). Furthermore, the exclusion zone will be `±2` (`np.ceil(6 / 4) = 2`). Thus the distance matrix will look something like this:

<img alt="Empty Distance Matrix" src="https://user-images.githubusercontent.com/35665983/131706193-ee2dcfda-96a2-4a17-87a2-f58567679d4b.png" width="500" />

The grey area is the exclusion zone, i.e. these are distances that we need not calculate.

`stump` processes each diagonal in the distance matrix separately, traversing allong the yellow dividers:

<img alt="stump Matrix" src="https://user-images.githubusercontent.com/35665983/131706282-9eceb223-5352-4e7a-9d6e-5831edbd69a7.png" width="500" />

The benefit with this method is the benefit we get from computing along diagonals. The distance value calculated in each point can be carried forward to calculate the distance value in the point that is diagonally below. For example, once we have computed the distance for `(0, 3)`, we may use that distance to efficiently calculate the distance for point `(1, 4)`, without having to explicitly calculate the distance for point `(1, 4)`. After that, we may use that to compute distance for point `(2, 5)`, then for `(3, 6)`, and so on.

While in practice this is very efficient, there are two issues with this.

1. Traversing one entire diagonal before starting the next diagonal makes the computation very cache inefficient
2. Reusing distance values to compute distances along the diagonal can cause us to lose precision of our distance values when the diagonals are significantly lengthy

# How `stump_tiles` works

The approach `stump_tiles` takes involves breaking down the distance matrix into smaller tiles, each of which will be divided into groups of diagonals.

Suppose the tile size is set to `4` and the number of diagonals to be iterated together is set to `2`. Then, the distance matrix will be divided like this:

<img alt="Tiles Matrix" src="https://user-images.githubusercontent.com/35665983/131706517-4d3e428f-8306-40e3-a0b8-e8a993eeb176.png" width="500" />

The matrix is split into `4 x 4` tiles (except near the edges, where the tiles are smaller). Tiles with blue borders indicate that there are distances to traverse within those tiles, while tiles with red borders indicate that we should skip those tiles as there are no distances to traverse within them. Further more, each of the diagonals within the blue tiles are split into groups of `2` (or less). Each diagonal groups are traversed together.

For example, if we were to traverse the tile shaded light pink, we would do so in the following order:

```
(5, 8)
(4, 8)
(6, 9)
(5, 9)
(7, 10)
(6, 10)
(7, 11)
(4, 9)
(4, 10)
(5, 10)
(5, 11)
(6, 11)
(4, 11)
```

# What this PR includes

## `config.STUMPY_MAX_TILE_LENGTH`

This defines the maximum tile length to be used, and is currently set to `4096`. In the example above, tile length was `4`. It is likely that the optimal value for this variable is platform-dependent.

## `config.STUMPY_N_DIAGONALS`

This defines the number of diagonals that should be traversed together within a tile, and is currently set to `64`. In the example above, number of diagonals that are traversed together was `2`. The optimal value for this variable is also likely platform-dependent.

## `core._get_tiles`

```
_get_tiles(m, n_A, n_B, diags, tile_length)

Split distance matrix into tiles of given size

Parameters
----------
m : int
    Window size

n_A : ndarray
    The length of time series `T_A`

n_B : ndarray
    The length of time series `T_B`

diags : ndarray
    The diagonal indices of interest

tile_length : int
    Maximum length of each tile

Returns
-------
tiles : ndarray
    An array of 7-element-arrays representing each tile:
    The first two values represent offset positions of the tile
    The second two values represent dimensions of the tile
    The third two values represent diagonal indices to traverse in the tile
    The seventh value represents number of distances in the tile
```

`_get_tiles` can be used to split the distance matrix into tiles of a given size. It returns `tiles`, an array of 7-element-arrays, with each 7-element-arrays representing one tile. The contents of each 7-element-array are [`y_offset`, `x_offset`, `tile_height`, `tile_width`, `tile_lower_diag`, `tile_upper_diag`, `tile_ndist`], in this specific order.

- `x_offset`, `y_offset`: position of the tile with respect to the distance matrix
- `tile_height`, `tile_width`: dimensions of the tile
- `tile_lower_diag`, `tile_upper_diag`: range of diagonals to be traversed within the tile
- `tile_ndist`: number of distances to be computed within the tile

The information provided by `_get_tiles` is used to split the distance matrix into tiles, and distribute the tiles into multiple threads.

## `stump_tiles`

The `stump_tiles` module is a modified version of `stump`. While stump calls `_stump` which calls `_compute_diagonal`, `stump_tiles` calls `_stump_tiles`, which calls `_compute_tiles`.

`_stump_tiles` will call `core._get_tiles` to get tiles data and pass each group of tiles to `_compute_tiles` in a single thread. `_compute_tiles` will process the given tiles and divide each tile into "chunks" of diagonals that are to be processed together.

## Testing

The following test cases have been added:

- `tests.test_stump_tiles` (replica of `tests.test_stump` but for `stump_tiles`)
- `tests.core.test_get_tiles`

Here's how to run these tests specifically:

```
pytest -x -W ignore::RuntimeWarning -W ignore::DeprecationWarning tests/test_stump_tiles.py tests/test_core.py::test_get_tiles
```

# Performance

`stump_tiles` performs poorly on relatively smaller datasets, but is significantly faster on larger ones:

|   `len(T_A)` |   `m` |   `stump` mean exec time |   `stump_tiles` mean exec time |   `stump` std error exec time |   `stump_tiles` std error exec time |   Percentage improvement | Results accurate   |
|-------------:|------:|-------------------------:|-------------------------------:|------------------------------:|------------------------------------:|-------------------------:|:-------------------|
|         1000 |    10 |               0.00498899 |                     0.00905045 |                   0.00012564  |                         0.00024308  |                -81.4085  | True               |
|         2500 |    10 |               0.0083944  |                     0.0289679  |                   0.000233611 |                         0.00027525  |               -245.086   | True               |
|         5000 |    10 |               0.0203253  |                     0.0676026  |                   0.00133888  |                         0.00101449  |               -232.604   | True               |
|        10000 |   100 |               0.052907   |                     0.287368   |                   0.00189274  |                         0.000715189 |               -443.157   | True               |
|       100000 |   100 |               5.10028    |                     5.59796    |                   0.0680912   |                         0.0683429   |                 -9.75783 | True               |
|       100000 |  5000 |               5.66884    |                     6.89013    |                   0.0397672   |                         0.0807766   |                -21.5438  | True               |
|       200000 |   100 |              32.1883     |                    22.1073     |                   0.0819303   |                         0.577238    |                 31.3189  | True               |
|       200000 |  5000 |              32.6387     |                    26.6489     |                   0.313708    |                         0.548431    |                 18.352   | True               |
|       250000 |   100 |              56.4471     |                    35.0073     |                   0.644344    |                         0.764895    |                 37.9821  | True               |
|       250000 |  5000 |              55.6449     |                    43.5155     |                   0.0835144   |                         0.734595    |                 21.798   | True               |

# Pull Request Checklist

- [x] Fork, clone, and checkout the newest version of the code
- [x] Create a new branch
- [x] Make necessary code changes
- [x] Install `black` (i.e., `python -m pip install black` or `conda install -c conda-forge black`)
- [x] Install `flake8` (i.e., `python -m pip install flake8` or `conda install -c conda-forge flake8`)
- [x] Install `pytest-cov` (i.e., `python -m pip install pytest-cov` or `conda install -c conda-forge pytest-cov`)
- [x] Run `black .` in the root stumpy directory
- [x] Run `flake8 .` in the root stumpy directory
- [x] Run `./setup.sh && ./test.sh` in the root stumpy directory
- [x] Reference a Github issue (and create one if one doesn't already exist)