@alvii147 2022-11-07
@seanlaw despite the success of #658, looks like `stump_tiles` still significantly outperforms `stump`! This is expected, because with #658, we saw the gap between `stump` and `stump_uint` get smaller and smaller with larger time series, whereas `stump_tiles` has always been better with larger time series. I also now have no doubt that we ARE in fact utilizing cache lines in `stump_tiles`, because `stump_tiles` does well despite the disadvantage of having to recalculate the distance for each tile.