@seanlaw 2022-11-08
@alvii147 Are you able to plot the performance for both `stumpy_uint64` and `stump_tiles` for a variety of `n`? I would like to see where `stump_tiles` overtake `stump_uint64`.

Also, considering the work in #592, I think we should allow that to be completed first, copy that `stump.py` file (and `aamp.py` file) into `stump_tiles`, and then modify it accordingly. I still think what you've done so far is great but it's still quite messy and hard to follow IMHO. How would you feel if we closed this PR and then individually cherry pick improvements and add them one at a time in separate PRs? And then, when we are happy with it, we'd have some simple heuristic that chooses `stump_uint64` for `n <= some_number` and uses `stump_tiles` for `n > some_number`.

What do you think? How would you like to proceed?