@alvii147 2022-11-08
> Also, considering the work in https://github.com/TDAmeritrade/stumpy/issues/592, I think we should allow that to be completed first, copy that stump.py file (and aamp.py file) into stump_tiles, and then modify it accordingly.

Sure that sounds good to me, ping me when that's merged and I can pull those changes onto this.

> I still think what you've done so far is great but it's still quite messy and hard to follow IMHO

I agree, I refactored a lot of things today and it's still hard to follow. The algorithm is pretty complicated and ugly to start with. When we pick this up again, I can look into making it more readable and adding comments. Besides, I think #535 should help once we're done with it.

> How would you feel if we closed this PR and then individually cherry pick improvements and add them one at a time in separate PRs?

I'm not really sure what you mean by cherry picking improvements. I don't know if we can break `stump_tiles` like that, it's more like one whole independent module.

> And then, when we are happy with it, we'd have some simple heuristic that chooses `stump_uint64` for `n <= some_number` and uses `stump_tiles` for `n > some_number`.

I agree with that. I'm okay with closing this for now.