@alvii147 2022-11-13
@seanlaw sure, here are the details of my machine:

```
Processor: 2.3 GHz 8-Core Intel Core i9
Graphics: Intel UHD Graphics 630 1536 MB
Memory: 64 GB 2667 MHz DDR4
OS: macOS Ventura 13.0
```

Here's the script I'm using to compute performance:

```python
import sys
import time
import warnings
from datetime import datetime
import subprocess
import numpy as np
import pandas as pd
from scipy.stats import ttest_ind
import stumpy


# run bash command
def run_command(cmd, bash_path):
    subprocess.run([bash_path, '-c', cmd], stdout=subprocess.PIPE)

# record given results dataframe to csv and md
# record given comment, src files, and diff between main and given branch
def record_results(df, bash_path, comment, branch_name, src_files):
    # create records directory
    records_dir = 'performance_records'
    now = datetime.now()
    dirname = records_dir + '/' + now.strftime('%Y-%m-%d-%H%M%S.%f')

    src_files_direname = f'{dirname}/src/'
    cmd = f'mkdir -p {src_files_direname}'
    run_command(cmd, bash_path)

    # record results to csv
    results_csv_filename = f'{dirname}/results.csv'
    with open(results_csv_filename, 'w+') as f:
        f.write(df.to_csv(index=False))

    # record results to md
    results_md_filename = f'{dirname}/results.md'
    with open(results_md_filename, 'w+') as f:
        f.write(df.to_markdown(index=False))

    # record comment
    comment_filename = f'{dirname}/comment.txt'
    with open(comment_filename, 'w+') as f:
        f.write(comment + '\n')

    # record branch diff against main
    diff_filename = f'{dirname}/diff.txt'
    cmd = f'git diff --color {branch_name}..main'
    cmd += ' >'
    cmd += ' ' + diff_filename
    run_command(cmd, bash_path)

    # record current state of src files
    for f in src_files:
        cmd = f'cp {f} {src_files_direname}'
        run_command(cmd, bash_path)

# run function and get execution time as well as function output
def timeit(func, n=3):
    times = []
    outputs = []
    for _ in range(n):
        time_elapsed = time.time()
        output = func()
        time_elapsed = time.time() - time_elapsed

        times.append(time_elapsed)
        outputs.append(output)

    return times, outputs

# compare performance of two functions
def compare_performance(control_func, experimental_func, control_func_name, experimental_func_name, compare_output_func, compile_first=False, n=3):
    if compile_first:
        control_func()
        experimental_func()

    # time function speeds
    control_func_times, control_func_outputs = timeit(control_func, n=n)
    experimental_func_times, experimental_func_outputs = timeit(experimental_func, n=n)

    # get mean times
    control_func_mean_t = np.mean(control_func_times)
    experimental_func_mean_t = np.mean(experimental_func_times)

    # perform t-test
    _, p_value = ttest_ind(control_func_times, experimental_func_times, equal_var=False)
    # get improvement percentage
    improvement = ((control_func_mean_t - experimental_func_mean_t) * 100) / control_func_mean_t

    print('---------------------------------------------------------')
    print('{0:<20} | {1} s'.format(control_func_name, round(control_func_mean_t, 5)))
    print('{0:<20} | {1} s'.format(experimental_func_name, round(experimental_func_mean_t, 5)))
    print('{0:<20} | {1}'.format('p-value', round(p_value, 5)))
    print('{0:<20} | {1}%'.format('improvement', round(improvement, 3)))
    print('---------------------------------------------------------\n')

    # check output accuracy
    accurate = True
    for o1 in control_func_outputs:
        if not compare_output_func(o1, experimental_func_outputs[0]):
            accurate = False
            break

    if accurate:
        for o2 in experimental_func_outputs:
            if not compare_output_func(o2, control_func_outputs[0]):
                accurate = False
                break

    # warn for inaccurate output
    if not accurate:
        warnings.warn('inaccurate results')

    return control_func_mean_t, experimental_func_mean_t, p_value, improvement, accurate

if __name__ == '__main__':
    CONTROL_FUNC_NAME = 'stump'
    EXPERIMENTAL_FUNC_NAME = 'stump_tiles'
    # len(T_A), m
    PARAMETERS = [
        (1000, 10),
        (2500, 10),
        (5000, 10),
        # (10000, 100),
        # (100000, 100),
        # (100000, 5000),
        # (200000, 100),
        # (200000, 5000),
        # (250000, 100),
        # (250000, 5000),
        # (300000, 100),
        # (300000, 5000),
    ]
    # top-k
    K = 3
    # number of times to run using each parameter configuration
    N = 5
    # should record results in performance_records/ directory
    RECORD_RESULTS = True
    # variables below are not used if RECORD_RESULTS = False
    # current branch name, used to record diffs
    BRANCH_NAME = 'alvii147/stump_tiles'
    # src files to record
    SRC_FILES = [
        'stumpy/stump_tiles.py',
        'stumpy/core.py',
        'stumpy/config.py'
    ]
    # path to bash exec
    BASH_PATH = '/bin/bash'

    argc = len(sys.argv)
    comment = ''
    if argc > 1:
        comment = sys.argv[1]

    compare_output_func = lambda a, b: np.allclose(np.array(a, dtype=np.float64), np.array(b, dtype=np.float64))
    df_columns = [
        '`len(T_A)`',
        '`m`',
        f'`{CONTROL_FUNC_NAME}` mean exec time',
        f'`{EXPERIMENTAL_FUNC_NAME}` mean exec time',
        't-test p-value',
        'Percentage improvement',
        'Results accurate',
    ]

    control_func = getattr(stumpy, CONTROL_FUNC_NAME)
    experimental_func = getattr(stumpy, EXPERIMENTAL_FUNC_NAME)
    compile_first = True
    table = []
    for params in PARAMETERS:
        T_A = np.random.rand(params[0])
        m = params[1]
        print(f'len(T_A) = {T_A.shape[0]}, m = {m}')

        control_func_mean_t, experimental_func_mean_t, p_value, improvement, accurate = compare_performance(
            control_func=lambda: control_func(T_A=T_A, m=m, k=K),
            experimental_func=lambda: experimental_func(T_A=T_A, m=m, k=K),
            control_func_name=CONTROL_FUNC_NAME,
            experimental_func_name=EXPERIMENTAL_FUNC_NAME,
            compare_output_func=compare_output_func,
            compile_first=compile_first,
            n=N,
        )
        row = [params[0], params[1], control_func_mean_t, experimental_func_mean_t, p_value, improvement, accurate]
        table.append(row)
        compile_first = False

    df = pd.DataFrame(
        table,
        columns=df_columns,
    )
    print('')
    print(df)
    if RECORD_RESULTS:
        record_results(df, BASH_PATH, comment, BRANCH_NAME, SRC_FILES)
```

Put this script in branch `alvii147/stump_tiles` and run it. Uncomment/add entries in `PARAMETERS` according to what your machine can handle. By default it'll just print out the results. If you want the results to also be stored in `csv` and `md` formats, set `RECORD_RESULTS = True`.

Note: you'll want to have `pandas` installed, and if you want to record results, you'll want `tabulate` as well.